+++ site/admin/pages/editor.php
@@ -5,9 +5,8 @@
   $edit = false;
   if(isset($_GET["SiteName"])){
     $edit = false;
-    $db_link = mysqli_connect(MYSQL_HOST,MYSQL_USER,MYSQL_PASSWORD,MYSQL_DATABASE);
     $sql = "SELECT * FROM Sites Where Name = '".$_GET["SiteName"]."'";
-    $db_erg = mysqli_query( $db_link, $sql );
+    $db_erg = mysqli_query( $U->$db_link, $sql );
     while ($zeile = mysqli_fetch_array( $db_erg, MYSQLI_ASSOC))
     {
       if($zeile["Name"] == $_GET["SiteName"]){