+++ site/login/2fa.php
@@ -13,15 +13,13 @@
   $g = new \Sonata\GoogleAuthenticator\GoogleAuthenticator();
   if(isset($_POST["secret"])&&isset($_POST["register"])){
     $secret = $_POST["secret"];
-    $db_link = mysqli_connect(MYSQL_HOST,MYSQL_USER,MYSQL_PASSWORD,MYSQL_DATABASE);
     $sql = "UPDATE User SET google_2fa='".$secret."' WHERE Username='".$_SESSION["User_Name"]."'";
-    $db_erg = mysqli_query( $db_link, $sql );
+    $db_erg = mysqli_query( $U->$db_link, $sql );
     echo str_replace("%a",$U->getLang("login.2fa.google_authenticator"),$U->getLang("login.2fa.succeed"));
   }elseif(isset($_POST["code"])&&isset($_POST["login"])){
     $code = $_POST["code"];
-    $db_link = mysqli_connect(MYSQL_HOST,MYSQL_USER,MYSQL_PASSWORD,MYSQL_DATABASE);
     $sql = "SELECT * FROM User WHERE Username='".$_SESSION['temp_User_Name']."';";
-    $db_erg = mysqli_query( $db_link, $sql );
+    $db_erg = mysqli_query( $U->$db_link, $sql );
     while ($zeile = mysqli_fetch_array( $db_erg, MYSQLI_ASSOC))
       {
         if($g->checkCode($zeile["google_2fa"], $code)){
@@ -32,10 +30,9 @@
         }
       }
   }elseif(isset($_POST["delete"])){
-    $db_link = mysqli_connect(MYSQL_HOST,MYSQL_USER,MYSQL_PASSWORD,MYSQL_DATABASE);
     echo $_SESSION['temp_User_Name'];
     $sql = "UPDATE User SET google_2fa='' WHERE Username='".$_SESSION['User_Name']."';";
-    $db_erg = mysqli_query( $db_link, $sql );
+    $db_erg = mysqli_query( $U->$db_link, $sql );
     header("Location: /index.php");
   }
 ?>