+++ site/login/login.php
@@ -20,9 +20,8 @@
   */
   $fa = false;
   //Make a mysql query to database User
-  $db_link = mysqli_connect(MYSQL_HOST,MYSQL_USER,MYSQL_PASSWORD,MYSQL_DATABASE);
   $sql = "SELECT * FROM User";
-  $db_erg = mysqli_query( $db_link, $sql );
+  $db_erg = mysqli_query( $U->$db_link, $sql );
   //True when the query comes from 2fa.php where the code is set
   if(isset($_SESSION["code"])){
     //Sets username to the username from 2fa.php