+++ site/profile.php
@@ -47,9 +47,8 @@
         ?>
         <br><a target="_blank" href="https://gravatar.com"><button><?php echo $U->getLang("profile.changePP"); ?></button></a><br />
         <?php
-        $db_link = mysqli_connect(MYSQL_HOST,MYSQL_USER,MYSQL_PASSWORD,MYSQL_DATABASE);
         $sql = "SELECT * FROM User WHERE Username='".$_SESSION["User_Name"]."'";
-        $db_erg = mysqli_query( $db_link, $sql );
+        $db_erg = mysqli_query( $U->$db_link, $sql );
         while ($zeile = mysqli_fetch_array( $db_erg, MYSQLI_ASSOC)){
           if($zeile["google_token"] =="" && file_exists("login/client_string.json")){
         ?>